# apps/api/security/openapi.json
{
  "components": {
    "schemas": {
...
      "RegistryUrlEnvelope": {
        "properties": {
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "VolumeSchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "host_path": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "owner_username": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "host_path",
          "status",
          "created_at"
        ],
        "type": "object"
      },
      "VolumesEnvelope": {
        "properties": {
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/VolumeSchema"
            },
            "type": "array"
          }
        },
        "required": [
          "volumes"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "OpenAPI spec for the 17 security-fuzzable endpoints (see .scratch/security-fuzzing). The 8 admin-gated operations are tagged `admin-gated` for the low-privilege RBAC-boundary pass. The spec is export-only and never served.",
    "license": {
      "name": ""
    },
    "title": "OpenWorkspace API — security fuzz surface",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/persistent-volumes": {
      "get": {
        "operationId": "list_persistent_volumes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumesEnvelope"
                }
              }
            },
            "description": "orphaned persistent volumes"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires can_manage_users or admin"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "summary": "The orphaned-volumes view (spec Decision 7 + user story 16): available only\nto system admins and `can_manage_users` holders, never scoped by group.\nOnly `orphaned` rows are returned; a still-referenced (`active`) volume is\nnot an orphan and stays hidden here.",
        "tags": [
          "admin-gated"
        ]
      }
    },
    "/api/registry": {
      "get": {
        "operationId": "get_registry",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "cached registry payload"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires can_manage_registry or admin"
          },
          "404": {
            "description": "no registry payload cached yet"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
    "/api/registry/url": {
      "get": {
        "operationId": "get_registry_url",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryUrlEnvelope"
                }
              }
            },
            "description": "configured registry url"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires can_manage_registry or admin"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
...
  }
}